+++ lib/get_connect/connect.dart
@@ -100,7 +100,9 @@ class GetConnect extends GetConnectInterface {
     this.maxAuthRetries = 1,
     this.allowAutoSignedCert = false,
     this.withCredentials = false,
-  });
+  }) {
+    $configureLifeCycle();
+  }
 
   bool allowAutoSignedCert;
   String userAgent;