+++ lib/get_utils/src/extensions/context_extensions.dart
@@ -92,7 +92,11 @@ extension ContextExtensionss on BuildContext {
   double get devicePixelRatio => MediaQuery.of(this).devicePixelRatio;
 
   /// similar to [MediaQuery.of(this).textScaleFactor]
+  // ignore: deprecated_member_use
   double get textScaleFactor => MediaQuery.of(this).textScaleFactor;
+
+  /// similar to [MediaQuery.of(this).textScaleFactor]
+  TextScaler get textScale => MediaQuery.of(this).textScaler;
 
   /// get the shortestSide from screen
   double get mediaQueryShortestSide => mediaQuerySize.shortestSide;