+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [4.7.3]
+
+Fix Snackbar in Flutter 3.38
+
+
+## [4.7.2]
+
+Fix PopScope widget
+
+## [4.7.1]
+
+Fix super parameters bug in GetConnect
+
 ## [4.7.0]
 
 Added wasm support to give people still using version 4 a longer lifespan. Version 5 is being moved to stable, but many people are still relying on version 4.